requirements.txt
django
djangorestframework
environs
drf_yasg